+++ .php-cs-fixer.php
@@ -2,6 +2,7 @@
 
 $finder = PhpCsFixer\Finder::create()
 	->in(__DIR__ . '/src')
+	->in(__DIR__ . '/generated') //TODO: maybe this is unnecessary???
 	->in(__DIR__ . '/build')
 	->in(__DIR__ . '/tests')
 	->in(__DIR__ . '/tools')
