+++ composer.json
@@ -39,7 +39,7 @@
       "pocketmine/bedrock-item-upgrade-schema": "~1.16.0+bedrock-1.21.110",
       "pocketmine/bedrock-protocol": "~54.0.0+bedrock-1.21.130",
       "pocketmine/binaryutils": "^0.2.1",
-      "pocketmine/callback-validator": "^1.0.2",
+      "pocketmine/callback-validator": "dev-master",
       "pocketmine/color": "^0.3.0",
       "pocketmine/errorhandler": "^0.7.0",
       "pocketmine/log": "^0.4.0",
@@ -63,7 +63,10 @@
    },
    "autoload": {
       "psr-4": {
-         "pocketmine\\": "src/"
+         "pocketmine\\": [
+            "src/",
+            "generated/"
+         ]
       },
       "files": [
          "src/CoreConstants.php"
@@ -94,7 +97,7 @@
          "@php build/generate-item-type-names.php vendor/pocketmine/bedrock-data/required_item_list.json",
          "@php build/generate-known-translation-apis.php",
          "@php build/generate-pocketmine-yml-property-consts.php",
-         "@php build/generate-registry-annotations.php src"
+         "@php build/generate-registry-interface.php src generated"
       ]
    }
 }
