+++ doxygen/doxygen.conf
@@ -875,6 +875,7 @@ WARN_LOGFILE           =
 # Note: If this tag is empty the current directory is searched.
 
 INPUT                  = ../src \
+                         ../generated \
                          ../vendor/pocketmine \
                          ./index.md
