+++ phpstan.neon.dist
@@ -34,6 +34,7 @@ parameters:
 	paths:
 		- build
 		- src
+		- generated
 		- tests/phpstan/DummyPluginOwned.php
 		- tests/phpstan/rules
 		- tests/phpunit
