+++ tests/phpstan/configs/phpstan-bugs.neon
@@ -25,138 +25,6 @@ parameters:
 			path: ../../../src/block/DoubleTallGrass.php
 
 		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:ACACIA_HANGING_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:ACACIA_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:BIRCH_HANGING_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:BIRCH_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:CHERRY_HANGING_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:CHERRY_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:CRIMSON_HANGING_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:CRIMSON_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:DARK_OAK_HANGING_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:DARK_OAK_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:JUNGLE_HANGING_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:JUNGLE_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:MANGROVE_HANGING_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:MANGROVE_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:OAK_HANGING_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:OAK_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:PALE_OAK_HANGING_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:PALE_OAK_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:SPRUCE_HANGING_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:SPRUCE_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:WARPED_HANGING_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
-			message: '#^Creating callable from a non\-native static method pocketmine\\item\\VanillaItems\:\:WARPED_SIGN\(\)\.$#'
-			identifier: callable.nonNativeMethod
-			count: 1
-			path: ../../../src/block/VanillaBlocks.php
-
-		-
 			message: '#^Strict comparison using \=\=\= between \*NEVER\* and 5 will always evaluate to false\.$#'
 			identifier: identical.alwaysFalse
 			count: 1
